server/requirements.txt
Flask
Werkzeug
flask-http-middleware
flask-cors
python-dotenv
FlightRadarAPI
Flask-Cors
mysql-connector-python-rf